[update] split cardList class
