+++ frontend-web/src/pages/Data/DataLayout.tsx
@@ -2,10 +2,12 @@ import React, { useEffect, useState } from 'react';
 import { Box, Grid } from '@mui/material';
 import TitleBox from '@/components/TitleBox';
 import AddIconButton from '@/components/button/AddIconButton';
-import CardList, { DataSetCard, DataSourceCard } from '@/components/CardList';
 import DatabaseService from '@/api/databaseService';
 import { STATUS } from '@/constant';
 import { useAlert } from 'react-alert';
+import { DatabaseCardList } from '@/components/list/DatabaseCardList';
+import { DatasetCardList } from '@/components/list/DatasetCardList';
+import CardList from '@/components/list/CardList';
 
 const DataLayout = props => {
   const { isViewMode, setDataSet } = props;
@@ -20,7 +22,6 @@ const DataLayout = props => {
   });
 
   const handleSelectDatabase = enteredData => {
-    console.log('handleUpdate', enteredData);
     return setSelectedDatabase(prevState => ({ ...prevState, ...enteredData }));
   };
 
@@ -37,7 +38,6 @@ const DataLayout = props => {
    */
   const getDatabaseList = () => {
     DatabaseService.selectDatabaseList().then(response => {
-      console.log('selectDatabaseList', response.data.data);
       setDatabaseList(response.data.data);
       if (response.data.data.length > 0) {
         setSelectedDatabase({ databaseId: response.data.data[0].id });
@@ -47,7 +47,6 @@ const DataLayout = props => {
 
   const getDatabaseInfo = () => {
     DatabaseService.selectDatabase(selectedDatabase.databaseId).then(response => {
-      console.log('getDatabaseInfo', response.data.data);
       setDatasetList(response.data.data.datasets);
       setTableList(response.data.data.tables);
     });
@@ -89,7 +88,7 @@ const DataLayout = props => {
       <Grid container spacing={5}>
         <Grid item xs={12} md={4}>
           <TitleBox title="데이터 소스" button={<AddIconButton link="source/create" />}>
-            <DataSourceCard
+            <DatabaseCardList
               data={databaseList}
               selectedData={selectedDatabase}
               disabledIcons={!!isViewMode}
@@ -103,7 +102,7 @@ const DataLayout = props => {
           <Grid container spacing={5}>
             <Grid item xs={12}>
               <TitleBox title="데이터 셋" button={<AddIconButton link={`set/create/${selectedDatabase.databaseId}`} />}>
-                <DataSetCard data={datasetList} selectDataset={selectDataset} disabledIcons={!!isViewMode} />
+                <DatasetCardList data={datasetList} selectDataset={selectDataset} disabledIcons={!!isViewMode} />
               </TitleBox>
             </Grid>
             <Grid item xs={12}>
